agent_forge/pyproject.toml
[project]
name = "eidos-e3"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6,<7",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = ["pytest>=8.4"]
